requirements.txt
Django~=4.2.5
httpx~=0.24.1
django-filter~=23.3
djangorestframework~=3.14.0
django_cleanup~=8.0.0
django-localflavor~=4.0
django-cors-headers~=4.2.0
Pillow~=10.0.1
cryptography~=36.0.2
requests~=2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.8
django-ckeditor==6.7.0